requirements.txt
Flask>=2.0.1
tensorflow-cpu==2.15.0
tensorflow-intel==2.15.0
keras==2.15.0
numpy>=1.23.5
pandas>=1.5.3
scikit-learn>=1.2.2
nltk>=3.8.1
gensim>=4.3.0
scipy>=1.10.1
protobuf>=4.21.12 